Finished Section 28

+++ utils/file.helper.js
@@ -2,10 +2,14 @@ const fs = require('fs');
 const path = require('path');
 
 exports.deleteFile = filePath => {
-    filePath = path.join(path.resolve('./'), filePath);
-    fs.unlink(filePath, err => {
-        if (err) {
-            throw new Error(err);
-        }
-    });
+    const currentPath = path.resolve('./');
+    filePath = path.join(currentPath, filePath);
+
+    if (fs.existsSync(filePath)) {
+        fs.unlink(filePath, err => {
+            if (err) {
+                throw new Error(err);
+            }
+        });
+    }
 };
